Switch X509CertGuard to db storage and urlencode
The X509CertGuard was requiring the user to perform newline stripping
from certificates, but this operation invalidates some certificates.
Therefore it is not possible to continue with this method.

This PR switches the expectation of certificate delivery to be
urlencoded and no longer with newlines striped.

This PR also stores the `X509CertGuard.ca_certificate` in the database
instead of on the filesystem.

This PR fully regenerates the migrations, which is a breaking change. It
comes with a .removal release note advising users as such.

https://pulp.plan.io/issues/6352
closes #6352