Adjust RHSM cert checking to only use base_path
The `RHSMCertGuard` was requiring the `settings.CONTENT_PATH_PREFIX` to
also be one of the authorized urls in the certificate. This is not
desirable due to compatability with a huge number of certificates
already deployed to clients.

This PR removes the `settings.CONTENT_PATH_PREFIX` from the
`request.path` so `RHSMCertGuard` authorization check. Now only the
`Distribution.base_path` is expected in the client RHSM certificate.

The test are also updated with new certificates and base_paths due to
authorized urls changing for RHSMCertGuard.

https://pulp.plan.io/issues/6694
closes #6694